Sph implemetation

+++ 3D-fluid-simulator/GraphicEngine/Shape/ShapeInstance/Point/Point.cpp
@@ -1,11 +1,15 @@
-#include <Point.h>
+#include "Point.h"
 
-Point::Point(Vertex center, uint64_t radius)
-	:Shape<GL_POINTS>(VertexArray({ center }, { 0 }), radius, false)
+#include "Essentials.h"
+#include "GL/glew.h"
+#include "Shape.h"
+
+Point::Point(Vertex center, float radius)
+    : Shape<GL_POINTS>(VertexArray({center}, {0}), radius, false)
 {
 }
 
 Essentials::DistributionShape Point::GetParticleDistribution()
 {
-	return Essentials::DistributionShape::SPHERE;
+  return Essentials::DistributionShape::SPHERE;
 }
